Добавление работы с БД: 1. Создание БД. 2. Добавление записей по городам при их первом выборе. 3. Обновление записей при их повторном выборе.
Не работает добавление и обновление. Выдаёт ошибку "E/SQLiteLog: (1) no such column:"
По дебагу выпадает на строке 89 WeatherDataSource.java.

+++ app/src/main/java/com/d/ivan/weatheredu/Model/CurrentWeatherMain.java
@@ -1,7 +1,7 @@
 package com.d.ivan.weatheredu.Model;
 
 public class CurrentWeatherMain {
-    public Double temp;
-    public long pressure;
-    public long humidity;
+    public double temp;
+    public double pressure;
+    public double humidity;
 }
